feat: поправил работу со снимками x-ray

+++ Drivers/ImageDisplayDriver.py
@@ -5,5 +5,5 @@ from PIL import Image
 class ImageDisplayDriver:
 
     @staticmethod
-    def save(data: SPDImage):
-        Image.fromarray(data.modified_image.astype(data.dtype)).save(data.folder + data.modified_name + data.extension)
+    def save(data: SPDImage, dtype=np.uint8):
+        Image.fromarray(data.modified_image.astype(dtype)).save(data.folder + data.modified_name + data.extension)
